backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
pydantic==2.12.0
python-multipart==0.0.20
motor==3.3.1
cloudinary==1.41.0
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.1.3
python-dotenv==1.1.1
email-validator==2.3.0
